pyproject.toml
[tool.poetry]
name = "api-cofradia"
version = "0.1.0"
description = ""
authors = ["eloyjm"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
unidecode = "^1.3.8"
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tensorflow-hub = "0.15"
pydantic = "^2.10.4"
python-decouple = "^3.8"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-app = "app.__main__:main"
